refactor: modify Post repository, services, controllers, etc

+++ src/modules/posts/services/DeletePostService.ts
@@ -1,20 +1,37 @@
+import { IUsersRepository } from '@modules/users/domain/repositories/IUsersRepository';
 import AppError from '@shared/errors/AppError';
 import { inject, injectable } from 'tsyringe';
 import { IDeletePost } from '../domain/models/IDeletePost';
-import { IPostsRepository } from '../domain/repositories/IPostsRepository';
+import { IPostsRepository } from '@modules/posts/domain/repositories/IPostsRepository';
 
 @injectable()
 class DeletePostService {
   constructor(
     @inject('PostsRepository')
     private postsRepository: IPostsRepository,
+
+    @inject('UsersRepository')
+    private usersRepository: IUsersRepository,
   ) {}
 
-  public async execute({ id }: IDeletePost): Promise<void> {
+  public async execute({ id, user_id }: IDeletePost): Promise<void> {
     const post = await this.postsRepository.findById(id);
 
     if (!post) {
-      throw new AppError('Post not found.');
+      throw new AppError('Post não existe.', 404);
+    }
+
+    const user = await this.usersRepository.findById(user_id);
+
+    if (!user) {
+      throw new AppError('Usuário não existe.');
+    }
+
+    console.log(post);
+
+    // user.id !== post.user.id
+    if (user.id !== post.id) {
+      throw new AppError('Usuário não autorizado', 401);
     }
 
     await this.postsRepository.remove(post);
